requirements.txt
azure-functions
xmltodict